+++ cli/__init__.py
@@ -6,7 +6,7 @@ import sys
 
 def _get_run_args(print_args: bool = True):
     from jina.logging import default_logger
-    from jina.parser import get_main_parser
+    from jina.parsers import get_main_parser
     from jina.helper import colored
 
     parser = get_main_parser()
