+++ cli/export.py
@@ -7,7 +7,7 @@ import os
 def api_to_dict():
     from jina.enums import BetterEnum
     from jina import __version__
-    from jina.parser import get_main_parser
+    from jina.parsers import get_main_parser
 
     from argparse import _StoreAction, _StoreTrueAction
     port_attr = ('help', 'choices', 'default', 'required', 'option_strings', 'dest')
