+++ tests/unit/docker/test_hub_list.py
@@ -1,5 +1,5 @@
 from jina.docker.hubio import HubIO
-from jina.parser import set_hub_list_parser
+from jina.parsers import set_hub_list_parser
 
 
 def test_hub_list_local_with_submodule():
