+++ tests/unit/flow/test_flow.py
@@ -10,7 +10,7 @@ from jina.checker import NetworkChecker
 from jina.enums import FlowOptimizeLevel, SocketType
 from jina.executors import BaseExecutor
 from jina.flow import Flow
-from jina.parser import set_pea_parser, set_ping_parser, set_pod_parser
+from jina.parsers import set_pea_parser, set_ping_parser, set_pod_parser
 from jina.peapods.pods import BasePod
 from jina.peapods.runtimes.local import LocalRuntime
 from jina.proto.jina_pb2 import DocumentProto
