+++ tests/unit/peapods/runtimes/remote/ssh/test_ssh_remote.py
@@ -2,7 +2,7 @@ import pytest
 
 from jina.enums import RemoteAccessType
 from jina.flow import Flow
-from jina.parser import set_pea_parser, set_pod_parser
+from jina.parsers import set_pea_parser, set_pod_parser
 from jina.peapods.pods import BasePod
 from jina.peapods.runtimes.ssh import SSHRuntime
 from jina.proto import jina_pb2
